refactor: player 패키지 분리
