+++ src/main/java/blackjack/model/player/Gamer.java
@@ -1,5 +1,8 @@
-package blackjack.model;
+package blackjack.model.player;
 
+import blackjack.model.Card;
+import blackjack.model.Cards;
+import blackjack.model.Score;
 import java.util.List;
 
 public class Gamer extends Player {
