+++ src/main/java/blackjack/model/player/Player.java
@@ -1,5 +1,8 @@
-package blackjack.model;
+package blackjack.model.player;
 
+import blackjack.model.Card;
+import blackjack.model.Cards;
+import blackjack.model.Score;
 import java.util.List;
 
 public abstract class Player {
